Fix systemic import errors and type conflicts across 60+ files
- Fix wrong Evaluator import (values.types.string → eval.runtime) in 36 files
- Fix wrong Arguments import (values.types.namespace → eval.runtime) in 15 files
- Fix wrong ParametersSpec import (eval.bc → eval.runtime.params.spec) in 6 files
- Fix wrong IrSpanned import (eval.compiler.args → eval.compiler) in 4 files
- Fix NativeFuncFn typedef referencing wrong Evaluator/Arguments types
- Rename custom Result<T,E> in typing/Ctx.kt to TypingResult to avoid
  conflict with kotlin.Result (eliminates 44 compile errors)
- Add GlobalsBuilder.setFunction convenience overload for simple lambdas
- Remove duplicate FrozenStringValue import in bc/compiler/Expr.kt

Reduces compile errors from ~10,021 to ~8,763 (~12.6% reduction)

+++ src/commonMain/kotlin/io/github/kotlinmania/starlark_kotlin/syntax/parser/Grammar.kt
@@ -1,11 +1,11 @@
 package io.github.kotlinmania.starlark_kotlin.syntax.parser
 
 object Grammar {
-    fun __action(state: Short, integer: Int): Short {
-        return GrammarState.ACTION[(state.toInt()) * 66 + integer]
+    fun __action(state: Int, integer: Int): Int {
+        return GrammarState.ACTION[state * 66 + integer].toInt()
     }
 
-    fun __goto(state: Short, nt: Int): Short {
+    fun __goto(state: Int, nt: Int): Int {
         
         return when (nt) {
             3 -> when (state) {
@@ -256,7 +256,7 @@ object Grammar {
         return 0
     }
 
-    fun __expected_tokens(__state: Short): List<String> {
+    fun __expected_tokens(__state: Int): List<String> {
         
         return emptyList()
     }
